python_sse/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
meraki==1.38.0
python-dotenv==1.0.0
httpx==0.25.2
pyyaml==6.0.1